Fix quest icon in Dialog
From @sevelev:

I found an issue that would occur when turning in quests that would ask you
to continue with the quest.  It would show the carbine "!" instead of the
one I made.  There was no call to it anywhere in the lua.  I discovered
that it was using the placeholder image for the ResponseItemIcon in the xml
file.  I changed that placeholder image to my "!".